requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distlib==0.4.0
Django==6.0.2
djangorestframework==3.16.1
fastapi==0.121.0
filelock==3.20.3
greenlet==3.2.4
h11==0.16.0
idna==3.11
Markdown==3.10.2
platformdirs==4.5.1
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.5
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.36.1
